refactor: 패키지 구조 변경 및 중복 파일 삭제 #23
