+++ app/src/main/java/com/boosters/promise/data/database/PromiseDatabase.kt
@@ -1,8 +1,9 @@
-package com.boosters.promise
+package com.boosters.promise.data.database
 
 import androidx.room.Database
 import androidx.room.RoomDatabase
 import androidx.room.TypeConverters
+import com.boosters.promise.data.promise.Promise
 
 @Database(entities = [Promise::class], version = 1)
 @TypeConverters(UserTypeConverter::class)
