+++ app/src/main/java/com/boosters/promise/data/network/PromiseService.kt
@@ -1,5 +1,6 @@
-package com.boosters.promise.network
+package com.boosters.promise.data.network
 
+import com.boosters.promise.data.promise.source.remote.PlaceResponse
 import retrofit2.http.GET
 import retrofit2.http.Query
 
@@ -9,5 +10,5 @@ interface PromiseService {
         @Query("query") query: String,
         @Query("display") display: Int? = null,
         @Query("start") start: Int? = null
-    ): LocalResponse
+    ): PlaceResponse
 }
