+++ app/src/main/java/com/boosters/promise/data/promise/di/InterfaceModule.kt
@@ -1,5 +1,11 @@
-package com.boosters.promise
+package com.boosters.promise.data.promise.di
 
+import com.boosters.promise.data.promise.source.PromiseRemoteDataSource
+import com.boosters.promise.data.promise.source.local.PromiseLocalDataSource
+import com.boosters.promise.data.promise.source.local.PromiseLocalDataSourceImpl
+import com.boosters.promise.data.promise.source.remote.PromiseRemoteDataSourceImpl
+import com.boosters.promise.data.promise.source.remote.PromiseRepository
+import com.boosters.promise.data.promise.source.remote.PromiseRepositoryImpl
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
