+++ app/src/main/java/com/boosters/promise/ui/promise/SearchAddressDialogFragment.kt
@@ -1,4 +1,4 @@
-package com.boosters.promise
+package com.boosters.promise.ui.promise
 
 import android.app.AlertDialog
 import android.app.Dialog
@@ -6,9 +6,12 @@ import android.os.Bundle
 import androidx.appcompat.widget.SearchView
 import androidx.databinding.DataBindingUtil
 import androidx.fragment.app.DialogFragment
+import com.boosters.promise.R
+import com.boosters.promise.data.network.Retrofit.promiseService
+import com.boosters.promise.data.promise.Place
 import com.boosters.promise.databinding.DialogSearchAddressBinding
-import com.boosters.promise.network.Retrofit.promiseService
-import com.boosters.promise.util.LocalMapper.asLocal
+import com.boosters.promise.ui.promise.adapter.SearchAddressListAdapter
+import com.boosters.promise.util.PlaceMapper.asPlace
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
@@ -17,10 +20,10 @@ import kotlinx.coroutines.withContext
 class SearchAddressDialogFragment : DialogFragment() {
 
     private lateinit var listener: SearchAddressDialogListener
-    private var searchResult: Local? = null
+    private var searchResult: Place? = null
 
     interface SearchAddressDialogListener {
-        fun onDialogPositiveClick(dialog: DialogFragment, resultItem: Local?)
+        fun onDialogPositiveClick(dialog: DialogFragment, resultItem: Place?)
         fun onDialogNegativeClick(dialog: DialogFragment)
     }
 
@@ -36,7 +39,7 @@ class SearchAddressDialogFragment : DialogFragment() {
                 DataBindingUtil.inflate(inflater, R.layout.dialog_search_address, null, false)
 
             val adapter = SearchAddressListAdapter { item ->
-                binding.searchViewDialogSearchAddress.setQuery(item.title, true)
+                binding.searchViewDialogSearchAddress.setQuery(item.title, false)
                 searchResult = item
             }
 
@@ -51,7 +54,7 @@ class SearchAddressDialogFragment : DialogFragment() {
                             val result = promiseService.searchLocalQuery(query.orEmpty(), 5)
                             withContext(Dispatchers.Main) {
                                 adapter.submitList(result.items.map { item ->
-                                    item?.asLocal()
+                                    item?.asPlace()
                                 })
                             }
                         }
