+++ app/src/main/java/com/boosters/promise/ui/promise/adapter/SearchAddressListAdapter.kt
@@ -1,4 +1,4 @@
-package com.boosters.promise
+package com.boosters.promise.ui.promise.adapter
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
@@ -6,11 +6,13 @@ import androidx.core.text.parseAsHtml
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
+import com.boosters.promise.R
+import com.boosters.promise.data.promise.Place
 import com.boosters.promise.databinding.ItemSearchAddressResultBinding
 
 class SearchAddressListAdapter(
-    private val onClickListener: (item: Local) -> Unit
-) : ListAdapter<Local, SearchAddressListAdapter.SearchAddressViewHolder>(diffUtil) {
+    private val onClickListener: (item: Place) -> Unit
+) : ListAdapter<Place, SearchAddressListAdapter.SearchAddressViewHolder>(diffUtil) {
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): SearchAddressViewHolder {
         return SearchAddressViewHolder(parent)
@@ -26,7 +28,7 @@ class SearchAddressListAdapter(
     ) {
         private val binding = ItemSearchAddressResultBinding.bind(itemView)
 
-        fun bind(item: Local) {
+        fun bind(item: Place) {
             binding.textViewItemSearchAddressResultTitle.text = item.title.parseAsHtml()
             itemView.setOnClickListener {
                 onClickListener(item)
@@ -35,12 +37,12 @@ class SearchAddressListAdapter(
     }
 
     companion object {
-        val diffUtil = object : DiffUtil.ItemCallback<Local>() {
-            override fun areItemsTheSame(oldItem: Local, newItem: Local): Boolean {
+        val diffUtil = object : DiffUtil.ItemCallback<Place>() {
+            override fun areItemsTheSame(oldItem: Place, newItem: Place): Boolean {
                 return oldItem.address == newItem.address
             }
 
-            override fun areContentsTheSame(oldItem: Local, newItem: Local): Boolean {
+            override fun areContentsTheSame(oldItem: Place, newItem: Place): Boolean {
                 return oldItem == newItem
             }
         }
