+++ app/src/main/java/com/boosters/promise/util/LocalMapper.kt
@@ -1,13 +1,13 @@
 package com.boosters.promise.util
 
 import androidx.core.text.parseAsHtml
-import com.boosters.promise.Local
-import com.boosters.promise.network.LocalItemResponse
+import com.boosters.promise.data.promise.Place
+import com.boosters.promise.data.promise.source.remote.PlaceItemResponse
 
-object LocalMapper {
-    fun LocalItemResponse.asLocal(): Local {
+object PlaceMapper {
+    fun PlaceItemResponse.asPlace(): Place {
         val title = this.title.parseAsHtml().toString()
-        return Local(
+        return Place(
             title = title,
             link = link,
             category = category,
